+++ data-es5.js
@@ -1,4 +1,8 @@
 require('object.assign').shim();
+var util = require('./data-common');
+
+var sparseNote = util.sparseNote;
+
 var temp = {};
 
 // exports browsers and tests
@@ -6,11 +10,6 @@ exports.name = 'ES5';
 exports.target_file = 'es5/index.html';
 exports.skeleton_file = 'es5/skeleton.html';
 
-var sparseNote = {
-  val: true,
-  note_id: 'sparse_arrays',
-  note_html: 'Internet Explorer 6 - 8 do not differentiate between a dense array with undefined values, and a sparse array. Specifically, `0 in [,]` and `0 in [undefined]` both yield false - whereas in a compliant browser, the former would give `false`, the latter `true`. As such, ES5 array iteration methods can only be shimmed reliably when dealing with dense arrays.'
-};
 exports.tests = [
 {
   name: 'Object/array literal extensions',
