+++ data-es6.js
@@ -1,30 +1,12 @@
 require('object.assign').shim();
+var util = require('./data-common');
+
+var flag = util.flag;
+var babel = util.babel;
+var typescript = util.typescript;
+var strict = util.strict;
 
 var temp = {};
-var flag = "flagged";
-var strict = "strict";
-var fallthrough = "needs-polyfill-or-native";
-
-var babel = {
-    regenerator: {
-        val: true,
-        note_id: "babel-regenerator",
-        note_html: "This feature requires <code>regenerator-runtime</code>, it's a part of <code>babel-polyfill</code> or <code>babel-runtime</code>."
-    }
-};
-
-var typescript = {
-    corejs: {
-        val: true,
-        note_id: "typescript-core-js",
-        note_html: "This feature is supported when using TypeScript with <a href='https://github.com/zloirock/core-js'>core-js</a>, or when a native ES6 host is used."
-    },
-    fallthrough: {
-        val: fallthrough,
-        note_id: "typescript-es6",
-        note_html: "TypeScript's compiler will accept code using this feature if the <code>--target ES6</code> flag is set, but passes it through unmodified and does not supply a runtime polyfill."
-    }
-};
 
 exports.name = 'ES6';
 exports.target_file = 'es6/index.html';
