+++ data-esintl.js
@@ -17,12 +17,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -35,12 +31,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
   ],
@@ -59,12 +51,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -78,12 +66,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -97,12 +81,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -121,9 +101,7 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -167,7 +145,6 @@ exports.tests = [
         safaritp: true,
         webkit: true,
         node012: true,
-        node4: true,
       },
     },
   ],
@@ -186,12 +163,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
   ],
@@ -210,12 +183,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
   ],
@@ -234,12 +203,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -252,12 +217,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -271,12 +232,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -290,12 +247,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -314,9 +267,7 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -360,7 +311,6 @@ exports.tests = [
         safaritp: true,
         webkit: true,
         node012: true,
-        node4: true,
       },
     },
   ],
@@ -379,12 +329,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -398,12 +344,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -417,12 +359,8 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -441,9 +379,7 @@ exports.tests = [
         edge12: true,
         firefox29: true,
         chrome24: true,
-        chrome29: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -487,7 +423,6 @@ exports.tests = [
         safaritp: true,
         webkit: true,
         node012: true,
-        node4: true,
       },
     },
     {
@@ -498,8 +433,6 @@ exports.tests = [
       */},
       res: {
         ie9: false,
-        ie10: false,
-        ie11: false,
         edge12: false,
         edge14: true,
         firefox29: false,
@@ -509,16 +442,11 @@ exports.tests = [
         safari51: null,
         safari6: null,
         safari7: false,
-        safari71_8: false,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
         opera12: null,
         android40: null,
         ios7: false,
-        ios9: false
       }
     },
     {
@@ -536,19 +464,13 @@ exports.tests = [
       */},
       res: {
         ie9: false,
-        ie10: false,
-        ie11: false,
         edge12: false,
         edge14: true,
         firefox29: false, // Firefox bug #1266290
         firefox52: true,
         chrome24: true,
-        chrome29: true,
         safari10: true,
-        safaritp: true,
-        webkit: true,
         node012: true,
-        node4: true,
       }
     }
   ],
@@ -564,27 +486,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
@@ -601,27 +509,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
@@ -638,27 +532,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
@@ -675,27 +555,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
@@ -712,27 +578,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
@@ -749,27 +601,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
@@ -786,27 +624,13 @@ exports.tests = [
       */},
       res: {
         ie9: true,
-        ie10: true,
-        ie11: true,
         edge12: true,
         firefox2: true,
-        firefox29: true,
         chrome22: true,
-        chrome24: true,
-        chrome29: true,
         safari51: true,
-        safari6: true,
-        safari7: true,
-        safari71_8: true,
-        safaritp: true,
-        webkit: true,
         opera12: true,
         ios7: true,
-        ios9: true,
         node010: true,
-        node012: true,
-        iojs: true,
-        node4: true,
         android40: true,
       },
     },
