+++ data-non-standard.js
@@ -9,39 +9,13 @@ exports.tests = [
   exec: function () {
     return typeof uneval == 'function';
   },
-  res: {
+  res: { 
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false,
@@ -63,37 +37,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: true,
     rhino17: true,
     phantom: false
@@ -107,31 +55,9 @@ exports.tests = [
   },
   res: {
     ie7: true,
-    ie11: true,
     iojs: true,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: true,
-    safari4: true,
-    safari5: true,
-    safari7: true,
-    safari71_8: true,
-    safaritp: true,
-    webkit: true,
     chrome7: true,
     opera10_10: false,
     opera10_50: true,
@@ -139,7 +65,6 @@ exports.tests = [
     chrome36: true,
     chrome49: true,
     konq44: true,
-    konq49: true,
     besen: false,
     rhino17: false,
     phantom: true,
@@ -155,7 +80,6 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -165,19 +89,9 @@ exports.tests = [
     firefox12: false,
     firefox28: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -193,24 +107,8 @@ exports.tests = [
   },
   res: {
     ie7: true,
-    ie11: true,
     iojs: true,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: null,
     safari4: true,
     safari5: true,
@@ -240,7 +138,6 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
     firefox3: false,
     firefox3_5: false,
@@ -250,27 +147,10 @@ exports.tests = [
     firefox12: true,
     firefox16: true,
     firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -286,7 +166,6 @@ exports.tests = [
     },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -296,19 +175,9 @@ exports.tests = [
     firefox12: false,
     firefox28: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -322,7 +191,6 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -332,19 +200,9 @@ exports.tests = [
     firefox12: false,
     firefox28: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -363,7 +221,6 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -379,19 +236,9 @@ exports.tests = [
     firefox37: true,
     firefox44: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -404,37 +251,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -447,37 +268,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -494,7 +289,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -511,19 +305,9 @@ exports.tests = [
     firefox44: true,
     firefox46: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -551,7 +335,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
     firefox3: true,
     firefox3_5: true,
@@ -561,27 +344,10 @@ exports.tests = [
     firefox12: true,
     firefox16: true,
     firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: true,
     rhino17: false,
     phantom: false
@@ -595,7 +361,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -605,19 +370,9 @@ exports.tests = [
     firefox12: true,
     firefox17: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -635,7 +390,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -647,19 +401,9 @@ exports.tests = [
     firefox17: true,
     firefox20: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -674,7 +418,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -726,37 +469,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -789,37 +506,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -861,7 +552,6 @@ exports.tests = [
   }],
   res: {
     ie7: false,
-    ie11: false,
     firefox3: true,
     firefox3_5: true,
     firefox4: true,
@@ -870,27 +560,10 @@ exports.tests = [
     firefox12: true,
     firefox16: true,
     firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -906,7 +579,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox3: true,
     firefox3_5: true,
     firefox4: true,
@@ -922,19 +594,9 @@ exports.tests = [
     firefox44: true,
     firefox46: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -977,32 +639,18 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
-    firefox3: false,
-    firefox3_5: false,
-    firefox4: false,
-    firefox5: false,
-    firefox7: false,
-    firefox12: false,
     firefox16: false,
     firefox17: false,
     firefox20: false,
     firefox28: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: true,
     opera10_50: true,
     chrome11: false,
     chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -1017,31 +665,9 @@ exports.tests = [
   },
   res: {
     ie7: true,
-    ie11: true,
     iojs: true,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: true,
-    safari4: true,
-    safari5: true,
-    safari7: true,
-    safari71_8: true,
-    safaritp: true,
-    webkit: true,
     chrome7: true,
     opera10_10: false,
     opera10_50: true,
@@ -1066,31 +692,9 @@ exports.tests = [
   },
   res: {
     ie7: true,
-    ie11: true,
     iojs: true,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: true,
-    safari4: true,
-    safari5: true,
-    safari7: true,
-    safari71_8: true,
-    safaritp: true,
-    webkit: true,
     chrome7: true,
     opera10_10: true,
     opera10_50: true,
@@ -1112,7 +716,6 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
     firefox3: false,
     firefox3_5: false,
@@ -1135,7 +738,6 @@ exports.tests = [
     chrome11: false,
     chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -1148,15 +750,7 @@ exports.tests = [
   */},
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
-    firefox3: false,
-    firefox3_5: false,
-    firefox4: false,
-    firefox5: false,
-    firefox7: false,
-    firefox12: false,
-    firefox28: false,
     safari3: null,
     safari4: false,
     safari5: false,
@@ -1183,7 +777,6 @@ exports.tests = [
   exec: function () { return typeof String.prototype.trimLeft === 'function' },
   res: {
     ie7: false,
-    ie11: false,
     edge13: true,
     iojs: true,
     firefox2: false,
@@ -1195,13 +788,6 @@ exports.tests = [
     firefox12: true,
     firefox16: true,
     firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
     safari4: false,
     safari5: true,
@@ -1211,7 +797,6 @@ exports.tests = [
     webkit: true,
     chrome7: true,
     opera10_10: false,
-    opera10_50: false,
     chrome11: true,
     chrome36: true,
     chrome49: true,
@@ -1229,7 +814,6 @@ exports.tests = [
   exec: function () { return typeof String.prototype.trimRight === 'function' },
   res: {
     ie7: false,
-    ie11: false,
     edge13: true,
     iojs: true,
     firefox2: false,
@@ -1241,13 +825,6 @@ exports.tests = [
     firefox12: true,
     firefox16: true,
     firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
     safari4: false,
     safari5: true,
@@ -1257,7 +834,6 @@ exports.tests = [
     webkit: true,
     chrome7: true,
     opera10_10: false,
-    opera10_50: false,
     chrome11: true,
     chrome36: true,
     chrome49: true,
@@ -1274,7 +850,6 @@ exports.tests = [
   exec: function () { return typeof String.prototype.quote === 'function' },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -1290,19 +865,9 @@ exports.tests = [
     firefox37: false,
     firefox44: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: null,
     rhino17: false,
     phantom: false
@@ -1314,7 +879,6 @@ exports.tests = [
   exec: function () { return 'foofoo'.replace('foo', 'bar', 'g') === 'barbar' },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
     firefox3: true,
     firefox3_5: true,
@@ -1358,37 +922,11 @@ exports.tests = [
   exec: function () { return typeof Date.prototype.toLocaleFormat === 'function' },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -1425,37 +963,11 @@ exports.tests = [
   exec: function () { return typeof Object.prototype.watch == 'function' },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -1467,37 +979,11 @@ exports.tests = [
   exec: function () { return typeof Object.prototype.unwatch == 'function' },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -1509,29 +995,11 @@ exports.tests = [
   exec: function () { return typeof Object.prototype.eval == 'function' },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
-    firefox3: false,
-    firefox3_5: false,
-    firefox4: false,
-    firefox5: false,
-    firefox7: false,
-    firefox12: false,
-    firefox28: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: true,
     rhino17: false,
     phantom: false
@@ -1547,8 +1015,7 @@ exports.tests = [
   res: {
     chrome36: true,
     chrome49: false,
-    node: true,
-    iojs: true,
+    node010: true,
     android50: true,
   },
   separator: 'after'
@@ -1568,36 +1035,12 @@ exports.tests = [
     ie11: true,
     iojs: true,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
     safari71_8: true,
-    safaritp: true,
-    webkit: true,
     chrome7: true,
     opera10_10: false,
     opera10_50: true,
-    chrome11: true,
-    chrome36: true,
-    chrome49: true,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false,
@@ -1612,37 +1055,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -1656,36 +1073,12 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: false,
-    firefox3: false,
-    firefox3_5: false,
-    firefox4: false,
-    firefox5: false,
-    firefox7: false,
-    firefox12: false,
     firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
@@ -1699,37 +1092,11 @@ exports.tests = [
   },
   res: {
     ie7: false,
-    ie11: false,
     firefox2: true,
-    firefox3: true,
-    firefox3_5: true,
-    firefox4: true,
-    firefox5: true,
-    firefox7: true,
-    firefox12: true,
-    firefox16: true,
-    firefox17: true,
-    firefox20: true,
-    firefox28: true,
-    firefox30: true,
-    firefox37: true,
-    firefox44: true,
-    firefox46: true,
-    firefox47: true,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: true,
     phantom: false
@@ -1743,29 +1110,11 @@ exports.tests = [
   },
   res: {
     ie7: true,
-    ie11: true,
     firefox2: false,
-    firefox3: false,
-    firefox3_5: false,
-    firefox4: false,
-    firefox5: false,
-    firefox7: false,
-    firefox12: false,
-    firefox28: false,
     safari3: false,
-    safari4: false,
-    safari5: false,
-    safari7: false,
-    safari71_8: false,
-    safaritp: false,
-    webkit: false,
     chrome7: false,
     opera10_10: false,
-    opera10_50: false,
-    chrome11: false,
-    chrome36: false,
     konq44: false,
-    konq49: false,
     besen: false,
     rhino17: false,
     phantom: false
